pyproject.toml
[tool.poetry]
name = "alexandria-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Lobo <matheuscardoso1@poli.ufrj.br>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.1"
python-telegram-bot = "^13.7"
requests = "^2.26.0"
bs4 = "^0.0.1"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
